+++ src/main/java/org/sonarsource/plugins/rtechnicaldebt/measures/cumulative/ComputeAverageMI.java
@@ -41,13 +41,13 @@ public class ComputeAverageMI implements MeasureComputer {
     // measure is already defined on files by {@link SetSizeOnFilesSensor}
     // in scanner stack
     if (context.getComponent().getType() != Component.Type.FILE) {
-      int sum = 0;
+      float sum = 0;
       int count = 0;
       for (Measure child : context.getChildrenMeasures(MARTINS_INSTABILITY.key())) {
         sum += child.getIntValue();
         count++;
       }
-      int average = count == 0 ? 0 : sum / count;
+      float average = count == 0 ? 0 : sum / count;
       context.addMeasure(MARTINS_INSTABILITY.key(), average);
     }
   }
